LPD-21824 Consolidate
